gemini_clasification_demo/requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0 